# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DiagnosticEnvelope": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "1.0.0",
            "title": "Api Version",
            "type": "string"
          },
          "axes": {
            "$ref": "#/components/schemas/DiagnosticAxes"
          },
          "consequences": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticConsequence"
            },
            "maxItems": 100,
            "title": "Consequences",
            "type": "array"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticLimitation"
            },
            "maxItems": 100,
            "title": "Limitations",
            "type": "array"
          },
          "provenance": {
            "$ref": "#/components/schemas/DiagnosticProvenance"
          },
          "repetitions": {
            "items": {
              "$ref": "#/components/schemas/RepetitionSummary"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Repetitions",
            "type": "array"
          },
          "request_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Request Fingerprint",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "selected_execution": {
            "$ref": "#/components/schemas/PreviewEnvelope"
          },
          "statistics": {
            "discriminator": {
              "mapping": {
                "DISTRIBUTION": "#/components/schemas/DistributionStatistics",
                "SINGLE_EXECUTION": "#/components/schemas/SingleExecutionStatistics"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/SingleExecutionStatistics"
              },
              {
                "$ref": "#/components/schemas/DistributionStatistics"
              }
            ],
            "title": "Statistics"
          }
        },
        "required": [
          "api_version",
          "schema_version",
          "job_id",
          "request_fingerprint",
          "statistics",
          "axes",
          "repetitions",
          "selected_execution",
          "consequences",
          "limitations",
          "provenance"
        ],
        "title": "DiagnosticEnvelope",
        "type": "object"
      },
...
      "DiagnosticRequest": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "1.0.0",
            "title": "Api Version",
            "type": "string"
          },
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          },
          "input_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Input Fingerprint",
            "type": "string"
          },
          "provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/OrigemValor"
            },
            "maxProperties": 500,
            "title": "Provenance",
            "type": "object"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "sampling": {
            "discriminator": {
              "mapping": {
                "FIXED_INPUT": "#/components/schemas/FixedInputPlan",
                "GENERATED_INPUT": "#/components/schemas/GeneratedInputPlan"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FixedInputPlan"
              },
              {
                "$ref": "#/components/schemas/GeneratedInputPlan"
              }
            ],
            "title": "Sampling"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_revision": {
            "minimum": 1.0,
            "title": "Scenario Revision",
            "type": "integer"
          },
          "selected_repetition_id": {
            "format": "uuid",
            "title": "Selected Repetition Id",
            "type": "string"
          },
          "study_id": {
            "format": "uuid",
            "title": "Study Id",
            "type": "string"
          }
        },
        "required": [
          "api_version",
          "request_id",
          "idempotency_key",
          "study_id",
          "scenario_id",
          "scenario_revision",
          "input_fingerprint",
          "sampling",
          "selected_repetition_id",
          "provenance"
        ],
        "title": "DiagnosticRequest",
        "type": "object"
      },
      "DiagnosticRetryRequest": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "1.0.0",
            "title": "Api Version",
            "type": "string"
          },
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "api_version",
          "request_id",
          "idempotency_key"
        ],
        "title": "DiagnosticRetryRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobSnapshot": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "1.0.0",
            "title": "Api Version",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "retry_of_job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of Job Id"
          },
          "status": {
            "enum": [
              "QUEUED",
              "RUNNING",
              "AGGREGATING",
              "CANCEL_REQUESTED",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "api_version",
          "job_id",
          "request_id",
          "status",
          "progress",
          "retry_of_job_id",
          "error"
        ],
        "title": "JobSnapshot",
        "type": "object"
      },
...
      "PreparationRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_build_sha": {
            "pattern": "^[0-9a-f]{40}$",
            "title": "Expected Build Sha",
            "type": "string"
          },
          "input": {
            "$ref": "#/components/schemas/EffectiveInput"
          },
          "preparation_version": {
            "const": "1.0.0",
            "title": "Preparation Version",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_revision": {
            "minimum": 1.0,
            "title": "Scenario Revision",
            "type": "integer"
          },
          "study_id": {
            "format": "uuid",
            "title": "Study Id",
            "type": "string"
          }
        },
        "required": [
          "preparation_version",
          "request_id",
          "study_id",
          "scenario_id",
          "scenario_revision",
          "expected_build_sha",
          "input"
        ],
        "title": "PreparationRequest",
        "type": "object"
      },
      "PreparationResponse": {
        "additionalProperties": false,
        "properties": {
          "composition": {
            "items": {
              "$ref": "#/components/schemas/RealizedComposition"
            },
            "title": "Composition",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "derived_provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DerivedEvidence"
            },
            "title": "Derived Provenance",
            "type": "object"
          },
          "generation_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Generation Fingerprint",
            "type": "string"
          },
          "generator_version": {
            "const": "dimensionamento-v1",
            "title": "Generator Version",
            "type": "string"
          },
          "input_snapshot": {
            "$ref": "#/components/schemas/EffectiveInput"
          },
          "motor_build_sha": {
            "pattern": "^[0-9a-f]{40}$",
            "title": "Motor Build Sha",
            "type": "string"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/OrdemEntrada"
            },
            "maxItems": 1000,
            "title": "Orders",
            "type": "array"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/PreparationParameter"
            },
            "title": "Parameters",
            "type": "array"
          },
          "preparation_id": {
            "format": "uuid",
            "title": "Preparation Id",
            "type": "string"
          },
          "preparation_version": {
            "const": "1.0.0",
            "title": "Preparation Version",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_revision": {
            "minimum": 1.0,
            "title": "Scenario Revision",
            "type": "integer"
          },
          "study_id": {
            "format": "uuid",
            "title": "Study Id",
            "type": "string"
          }
        },
        "required": [
          "preparation_version",
          "preparation_id",
          "request_id",
          "study_id",
          "scenario_id",
          "scenario_revision",
          "created_at",
          "motor_build_sha",
          "generator_version",
          "generation_fingerprint",
          "input_snapshot",
          "orders",
          "parameters",
          "composition",
          "derived_provenance"
        ],
        "title": "PreparationResponse",
        "type": "object"
      },
...
      "PreviaRequest": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "1.0.0",
            "title": "Api Version",
            "type": "string"
          },
          "cenario": {
            "$ref": "#/components/schemas/CenarioEntrada"
          },
          "periodo": {
            "discriminator": {
              "mapping": {
                "LEGADO": "#/components/schemas/PeriodoLegado",
                "NATURAL": "#/components/schemas/PeriodoNatural"
              },
              "propertyName": "modo"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PeriodoLegado"
              },
              {
                "$ref": "#/components/schemas/PeriodoNatural"
              }
            ],
            "title": "Periodo"
          },
          "proveniencia": {
            "additionalProperties": {
              "$ref": "#/components/schemas/OrigemValor"
            },
            "title": "Proveniencia",
            "type": "object"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_revision": {
            "minimum": 1.0,
            "title": "Scenario Revision",
            "type": "integer"
          },
          "study_id": {
            "format": "uuid",
            "title": "Study Id",
            "type": "string"
          }
        },
        "required": [
          "api_version",
          "request_id",
          "study_id",
          "scenario_id",
          "scenario_revision",
          "cenario",
          "periodo",
          "proveniencia"
        ],
        "title": "PreviaRequest",
        "type": "object"
      },
...
      "ReferenceExample": {
        "additionalProperties": false,
        "properties": {
          "cenario": {
            "$ref": "#/components/schemas/CenarioEntrada"
          },
          "periodo": {
            "discriminator": {
              "mapping": {
                "LEGADO": "#/components/schemas/PeriodoLegado",
                "NATURAL": "#/components/schemas/PeriodoNatural"
              },
              "propertyName": "modo"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PeriodoLegado"
              },
              {
                "$ref": "#/components/schemas/PeriodoNatural"
              }
            ],
            "title": "Periodo"
          },
          "proveniencia": {
            "additionalProperties": {
              "$ref": "#/components/schemas/OrigemValor"
            },
            "title": "Proveniencia",
            "type": "object"
          }
        },
        "required": [
          "cenario",
          "periodo",
          "proveniencia"
        ],
        "title": "ReferenceExample",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Motor de Fluxo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/diagnosticos": {
      "post": {
        "operationId": "diagnostic_schema_api_v1_diagnosticos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Schema"
      }
    },
    "/api/v1/diagnosticos/{job_id}": {
      "get": {
        "operationId": "diagnostic_job_schema_api_v1_diagnosticos__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Job Schema"
      }
    },
    "/api/v1/diagnosticos/{job_id}/cancelamentos": {
      "post": {
        "operationId": "diagnostic_cancellation_schema_api_v1_diagnosticos__job_id__cancelamentos_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Cancellation Schema"
      }
    },
    "/api/v1/diagnosticos/{job_id}/resultado": {
      "get": {
        "operationId": "diagnostic_result_schema_api_v1_diagnosticos__job_id__resultado_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Result Schema"
      }
    },
    "/api/v1/diagnosticos/{job_id}/retries": {
      "post": {
        "operationId": "diagnostic_retry_schema_api_v1_diagnosticos__job_id__retries_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticRetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Retry Schema"
      }
    },
    "/api/v1/examples/reference": {
      "get": {
        "operationId": "reference_example_schema_api_v1_examples_reference_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceExample"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reference Example Schema"
      }
    },
...
    "/api/v1/preparacoes": {
      "post": {
        "operationId": "preparation_api_v1_preparacoes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreparationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preparation"
      }
    },
    "/api/v1/previas": {
      "post": {
        "operationId": "preview_schema_api_v1_previas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Schema"
      }
    },
    "/api/v1/replays": {
      "post": {
        "operationId": "replay_schema_api_v1_replays_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayDocumentV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replay Schema"
      }
    },
    "/api/v1/session": {
      "get": {
        "operationId": "session_schema_api_v1_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Session Schema"
      }
    }
  }
}